django-prj/requirements.txt
Django==5.0.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==20.1.0
